refactor: favorites/route.ts を handlers/ と services/ に分解 #8
